Add files via upload
 Imprint Impact - Book Donation Management System  

Welcome to I"mprint Impact" – a platform designed to simplify and streamline book donations! This system connects donors and recipients, making it easier for people to share knowledge and resources through books. Whether you're donating your pre-loved books or looking to support education initiatives, Imprint Impact provides an efficient and user-friendly way to make a difference.  

  Key Features:  
- Easy Book Donations: Seamlessly register and donate books online.  
- Search & Request Books : Browse the collection and request books you need.  
- Admin Controls : Manage donations, users, and requests with ease.  

This project uses PHP and MySQL for backend development and is hosted locally with XAMPP.